refactor: 更改 term 模块软件包名
