+++ term/src/main/AndroidManifest.xml
@@ -15,12 +15,12 @@
 
     <application>
         <activity
-            android:name=".TermPreferences"
+            android:name="yangfentuozi.runner.term.TermPreferences"
             android:label="@string/preferences"
             android:theme="@style/AppTheme" />
 
         <service
-            android:name=".TermService"
+            android:name="yangfentuozi.runner.term.TermService"
             android:exported="false"
             android:foregroundServiceType="dataSync">
             <intent-filter>
