+++ term/src/main/java/yangfentuozi/runner/term/WindowListCloseButton.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package yangfentuozi.runner.app.ui.fragment.terminal;
+package yangfentuozi.runner.term;
 
 import android.content.Context;
 import android.util.AttributeSet;
