+++ term/src/main/res/layout/term_activity.xml
@@ -25,7 +25,7 @@
         android:layout_height="0dp"
         android:layout_weight="1">
 
-        <yangfentuozi.runner.app.ui.fragment.terminal.TermViewFlipper
+        <yangfentuozi.runner.term.TermViewFlipper
             android:id="@+id/view_flipper"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
